Add action in lieu of suspend
