+++ mod.ts
@@ -6,3 +6,4 @@ export * from "./spawn.ts";
 export * from "./run.ts";
 export * from "./context.ts";
 export * from "./resource.ts";
+export * from "./action.ts";
