+++ sleep.ts
@@ -1,8 +1,8 @@
-import { suspend } from "./suspend.ts";
+import { action } from "./action.ts";
 import type { Operation } from "./types.ts";
 
 export function sleep(duration: number): Operation<void> {
-  return suspend((resolve) => {
+  return action((resolve) => {
     let timeoutId = setTimeout(resolve, duration);
     return () => {
       clearTimeout(timeoutId);
