+++ suspend.ts
@@ -1,40 +1,10 @@
-import { Do, Resume } from "./control.ts";
-import { Err, Ok, Result } from "./result.ts";
-import type { Operation, Reject, Resolve } from "./types.ts";
+import type { Operation } from "./types.ts";
+import { Do } from "./control.ts";
 
-export function suspend(): Operation<void>;
-export function suspend<T>(resume: Resolver<T>): Operation<T>;
-export function suspend(unsuspend?: Resolver<unknown>): Operation<unknown> {
+export function suspend(): Operation<void> {
   return {
     *[Symbol.iterator]() {
-      let exit = () => {};
-      try {
-        let value = yield Do(({ next }) => {
-          if (unsuspend) {
-            let settlement: Result<unknown> | void = void 0;
-            let settle = (result: Result<unknown>) => {
-              if (!settlement) {
-                settlement = result;
-                next(Resume(settlement));
-              }
-            };
-            let resolve = (value: unknown) => settle(Ok(value));
-            let reject = (error: Error) => settle(Err(error));
-            try {
-              exit = unsuspend(resolve, reject) ?? (() => {});
-            } catch (error) {
-              next(Resume(Err(error)));
-            }
-          }
-        });
-        return value;
-      } finally {
-        exit();
-      }
+      yield Do(() => {});
     },
   };
-}
-
-export interface Resolver<T> {
-  (resolve: Resolve<T>, reject: Reject): () => void;
 }
